requirements.txt
aiohttp
aiowiki @ git+https://github.com/Gelbpunkt/aiowiki
async-cse
dateparser
discordemoji
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus
discord.py
humanize
jishaku
motor
Pillow
psutil
pytimeparse
PyNaCl
pomice @ git+https://github.com/Tari-dev/pomice
python-dotenv
qrcode
selenium
tableascii
uvloop
